refactor: 상수 변경
